test(migration): tidy migration tests, drop redundant cases
- Merge the two near-identical 20260606 tests (Idempotent +
  PartialCustomization) into one TestMigrate20260606_PreservesUserOverride;
  keep a focused TestMigrate20260606_DefaultsXcodeSkipUsage for the default.
- Organize by concern: migration_test.go holds all migration-function
  tests (ordered by date); migration_helpers_test.go holds only the helper
  unit tests plus shared svc()/countRules() fixtures.
- Replace ad-hoc duplicate-counting loops with the countRules() helper.

No production code change; config suite passes.
